+++ api/src/Communication/Service/NotificationManager.php
@@ -414,8 +414,8 @@ class NotificationManager
                 case Review::class:
                     $titleContext = [];
                     $bodyContext = [
-                        'givenName'=>$recipient->getGivenName(),
-                        'shortFamilyName'=>$recipient->getShortFamilyName()
+                        'givenName'=>$object->getReviewer()->getGivenName(),
+                        'shortFamilyName'=>$object->getReviewer()->getShortFamilyName()
                     ];
                     break;
                 default:
@@ -578,8 +578,8 @@ class NotificationManager
                     break;
                 case Review::class:
                     $bodyContext = [
-                        'givenName'=>$recipient->getGivenName(),
-                        'shortFamilyName'=>$recipient->getShortFamilyName()
+                        'givenName'=>$object->getReviewer()->getGivenName(),
+                        'shortFamilyName'=>$object->getReviewer()->getShortFamilyName()
                     ];
                     break;
                 default:
@@ -722,7 +722,7 @@ class NotificationManager
                     $bodyContext = ['user'=>$recipient];
                     break;
                 case Message::class:
-                    $titleContext = [];
+                    $titleContext = ['user'=>$object->getUser()];
                     $bodyContext = ['text'=>$object->getText(), 'user'=>$recipient];
                     break;
                 case CarpoolItem::class:
@@ -756,8 +756,8 @@ class NotificationManager
                 case Review::class:
                     $titleContext = [];
                     $bodyContext = [
-                        'givenName'=>$recipient->getGivenName(),
-                        'shortFamilyName'=>$recipient->getShortFamilyName()
+                        'givenName'=>$object->getReviewer()->getGivenName(),
+                        'shortFamilyName'=>$object->getReviewer()->getShortFamilyName()
                     ];
                     break;
                 default: