+++ api/src/User/Entity/User.php
@@ -104,6 +104,7 @@ use App\Payment\Ressource\BankAccount;
 use App\Solidary\Entity\Operate;
 use App\Solidary\Entity\SolidaryUser;
 use App\User\Controller\UserCanUseEmail;
+use App\User\Controller\UserSendValidationEmail;
 
 /**
  * A user.
@@ -300,6 +301,12 @@ use App\User\Controller\UserCanUseEmail;
  *              "method"="GET",
  *              "path"="/users/{id}/generate_phone_token",
  *              "controller"=UserGeneratePhoneToken::class,
+ *              "security"="is_granted('user_update',object)"
+ *          },
+ *          "send_validation_email"={
+ *              "method"="GET",
+ *              "path"="/users/{id}/sendValidationEmail",
+ *              "controller"=UserSendValidationEmail::class,
  *              "security"="is_granted('user_update',object)"
  *          },
  *          "alerts"={
@@ -503,6 +510,12 @@ class User implements UserInterface, EquatableInterface
      * @Groups({"readUser","write","checkValidationToken","passwordUpdateRequest","passwordUpdate", "readSolidary"})
      */
     private $email;
+
+    /**
+     * @var string|null The email of the user.
+     * @Groups({"readUser", "write"})
+     */
+    private $oldEmail;
 
     /**
      * @var string The email of the user in a professional context.
@@ -1303,14 +1316,26 @@ class User implements UserInterface, EquatableInterface
         return $this;
     }
 
-    public function getEmail(): string
+    public function getEmail(): ?string
     {
         return $this->email;
     }
 
-    public function setEmail(string $email): self
+    public function setEmail(?string $email): self
     {
         $this->email = $email;
+
+        return $this;
+    }
+
+    public function getOldEmail(): ?string
+    {
+        return $this->oldEmail;
+    }
+
+    public function setOldEmail(?string $oldEmail): self
+    {
+        $this->oldEmail = $oldEmail;
 
         return $this;
     }
@@ -1568,7 +1593,7 @@ class User implements UserInterface, EquatableInterface
     public function setPwdToken(?string $pwdToken): self
     {
         $this->pwdToken = $pwdToken;
-        $this->setPwdTokenDate($pwdToken ? new DateTime() : null);
+        $this->setPwdTokenDate($pwdToken ? new \DateTime() : null);
         return $this;
     }